reorganize database structure
