+++ backend/backend.py
@@ -1,25 +1,35 @@
-from app import create_app
+# from app import create_app
+from pathlib import Path
+from flask import Flask, Blueprint
 from flask_cors import CORS
-from routes.Database import config
+from flask_sqlalchemy import SQLAlchemy
 
-app = create_app()
-# app = Flask(__name__)
+
+basedir = Path(__file__).resolve().parent
+app = Flask(__name__)
+
+
 # set db paths
-app.config['SQLALCHEMY_DATABASE_URI'] = config.SQLALCHEMY_DATABASE_URI_1
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = config.SQLALCHEMY_TRACK_MODIFICATIONS_1
+DATABASE = "uevent.db"
+SQLALCHEMY_DATABASE_URI_1 = "sqlite:///" + str(Path(basedir).joinpath(DATABASE))  # noqa
+SQLALCHEMY_TRACK_MODIFICATIONS_1 = False
+app.config['SQLALCHEMY_DATABASE_URI'] = SQLALCHEMY_DATABASE_URI_1
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = SQLALCHEMY_TRACK_MODIFICATIONS_1
+
 
 # initialize DB
-from routes.Database.Event_DB import t_event_filter_db
-t_event_filter_db.init_app(app)
-
-with app.app_context():
-    # create the database and the db table
-    t_event_filter_db.create_all()
-
-    # commit the changes
-    t_event_filter_db.session.commit()
-
+db = SQLAlchemy(app)
+# initialize route path
+user = Blueprint("user", "user_authentication_system.py", url_prefix="/user")
+main_sys = Blueprint("main_sys", "main_system.py", url_prefix="/main_sys")
+detail = Blueprint("detail", "detailed_system.py", url_prefix="/detail")
+app.register_blueprint(user)
+app.register_blueprint(main_sys)
+app.register_blueprint(detail)
+# cors add
 CORS(app, resources={r"/user/*": {"origins": "http://localhost:3000"}})
 CORS(app, resources={r"/main_sys/*": {"origins": "http://localhost:3000"}})
+
+
 if __name__ == "__main__":
     app.run(debug=True)
