+++ backend/create_db.py
@@ -1,0 +1,7 @@
+from backend import *
+
+with app.app_context():
+    # create the database and the db table
+    db.create_all()
+    # commit the changes
+    db.session.commit()
