+++ backend/routes/detailed_system.py
@@ -1,13 +1,14 @@
 from flask import Blueprint, jsonify, render_template, request, redirect, url_for
+from backend import detail
 
-detail = Blueprint("detail",__name__, url_prefix="/detail")
 
 @detail.route("/", methods=['POST'])
 def display():
     # page not built yet, return to index with nothing for now
     return render_template("index.html")
 
-@detail.route("/add", methods=['POST','GET'])
+
+@detail.route("/add", methods=['POST', 'GET'])
 def add_entry():
     """Adds new comment to the comment section."""
     data = request.get_json()
@@ -16,9 +17,5 @@ def add_entry():
 
     if (comment_title) and (comment_content):
         return redirect(url_for("detail.display"))
-    
+
     return jsonify({'message': 'Empty field not allowed'}), 406
-    
-    
-    
-
