+++ backend/routes/user_authentication_system.py
@@ -1,15 +1,15 @@
 from flask import Blueprint, jsonify, render_template, request
+from backend import user
+# API Route
 
-user = Blueprint("user",__name__, url_prefix="/user")
-## API Route
 
-@user.route("/login", methods=['POST','GET'])
+@user.route("/login", methods=['POST'])
 def login():
     data = request.get_json()
     username = data.get('username')
     password = data.get('password')
 
     if not username or not password:
-        return jsonify({'message': 'One of the entries is empty'}), 401
+        return jsonify({'message': 'One of the entries is empty'}), 406
     else:
         return jsonify({'message': 'Login successful'}), 200
